+++ extensions/machine-learning-services/src/views/models/modelManagementController.ts
@@ -9,7 +9,8 @@ import { azureResource } from '../../typings/azure-resource';
 import { ApiWrapper } from '../../common/apiWrapper';
 import { AzureModelRegistryService } from '../../modelManagement/azureModelRegistryService';
 import { Workspace } from '@azure/arm-machinelearningservices/esm/models';
-import { RegisteredModel, WorkspaceModel, PredictParameters, DatabaseTable } from '../../modelManagement/interfaces';
+import { RegisteredModel, WorkspaceModel } from '../../modelManagement/interfaces';
+import { PredictParameters, DatabaseTable } from '../../prediction/interfaces';
 import { RegisteredModelService } from '../../modelManagement/registeredModelService';
 import { RegisteredModelsDialog } from './registerModels/registeredModelsDialog';
 import { AzureResourceEventArgs, ListAzureModelsEventName, ListSubscriptionsEventName, ListModelsEventName, ListWorkspacesEventName, ListGroupsEventName, ListAccountsEventName, RegisterLocalModelEventName, RegisterLocalModelEventArgs, RegisterAzureModelEventName, RegisterAzureModelEventArgs, ModelViewBase, SourceModelSelectedEventName, RegisterModelEventName, DownloadAzureModelEventName, ListDatabaseNamesEventName, ListTableNamesEventName, ListColumnNamesEventName, PredictModelEventName, PredictModelEventArgs } from './modelViewBase';
